Sonic Hose + UI Upgrades!
Sonic Hose is in action!

Animation for equipping, using, idling, and putting it away. There's
even a battery meter in the UI to show how long you can use it, audio
to indicate when it's in action along with a visual distortion field!
It recharges itself, and there's even a restriction on initiation when
it's below a certain charge.

UI for Sonic Hose, K-Bombs, and the Black Death are up and running!

I also reorganized the SonicHose UI, K-Bomb UI, and Black Death UI
scripts with inheritance. Similarly I did this with the Inventory, Gear
Selector, and Gear User scripts.

Lastly, I was able to create some custom inspectors for the UI and Gear
Inventory scripts for optimizing workflow.